Easy Sql: Fixes for bugs and errors found by Tests (made by Copilot)

+++ sql-macros/src/sql_macros_components/sql_limit.rs
@@ -39,7 +39,7 @@ impl SqlLimit {
         match self {
             SqlLimit::Literal(l) => {
                 quote! {#sql_crate::LimitClause{
-                    limit: #l,
+                    limit: #l as usize,
                 }}
             }
             SqlLimit::Expr(expr) => {
